[FEAT] #115 신고 request 모델 수정

+++ app/src/main/java/com/starters/hsge/data/model/remote/request/ReportRequest.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ReportRequest(
     @SerialName("description") val description : String,
-    @SerialName("reportee") val reportee : Int
+    @SerialName("reportee") val reportee : Long
 )
